prism: split switch.go TUI/project-discovery/dashboard parts (#1158)
Extract safe-to-extract parts of cmd/switch.go (1230 LoC) into three
focused files per D2 §7:

- switch_tui.go: fuzzy pickerModel, inputModel, fuzzyMatch, pick,
  pickString, promptInput, promptBranchInput
- switch_project.go: entry type, expandHome, switchWorktreeExcludeSet,
  switchProjectLocations, switchProjectSpecific, projectEntries,
  handleBareRepo, handleRegularRepo, handleCloneRepo,
  activeReviewSessionEntries, handleReviewGroupPick
- switch_dashboard.go: ensureSwitchDashSession

switch.go retains only session-management (injectContainerConfig,
ensureAndSwitch, allocatePortForSession) and the cobra command — the
switch_session.go extraction remains deferred until A.1 T2-8 lands.

No behaviour changes; all functions moved verbatim with imports adjusted.
